+++ std/recursion/plonk/batching_circuit_test.go
@@ -1,7 +1,6 @@
 package plonk
 
 import (
-	"fmt"
 	"math/big"
 	"testing"
 
@@ -15,10 +14,7 @@ import (
 	"github.com/consensys/gnark/constraint"
 	"github.com/consensys/gnark/frontend"
 	"github.com/consensys/gnark/frontend/cs/scs"
-	"github.com/consensys/gnark/std/algebra"
 	"github.com/consensys/gnark/std/algebra/native/sw_bls12377"
-	gnark_kzg "github.com/consensys/gnark/std/commitments/kzg"
-	"github.com/consensys/gnark/std/math/emulated"
 	"github.com/consensys/gnark/std/recursion"
 	"github.com/consensys/gnark/test"
 )
@@ -27,12 +23,12 @@ import (
 // inner circuits
 
 // inner circuit
-type InnerCircuit struct {
+type InnerCircuit1 struct {
 	X frontend.Variable
 	Y frontend.Variable `gnark:",public"`
 }
 
-func (c *InnerCircuit) Define(api frontend.API) error {
+func (c *InnerCircuit1) Define(api frontend.API) error {
 	var res frontend.Variable
 	res = c.X
 	for i := 0; i < 5; i++ {
@@ -51,9 +47,9 @@ func (c *InnerCircuit) Define(api frontend.API) error {
 }
 
 // get VK, PK base circuit
-func GetInnerCircuitData() (constraint.ConstraintSystem, native_plonk.VerifyingKey, native_plonk.ProvingKey, kzg.SRS) {
+func getInnerCircuitData() (constraint.ConstraintSystem, native_plonk.VerifyingKey, native_plonk.ProvingKey, kzg.SRS) {
 
-	var ic InnerCircuit
+	var ic InnerCircuit1
 	ccs, err := frontend.Compile(ecc.BLS12_377.ScalarField(), scs.NewBuilder, &ic)
 	if err != nil {
 		panic("compilation failed: " + err.Error())
@@ -79,7 +75,7 @@ func getProofsWitnesses(assert *test.Assert, ccs constraint.ConstraintSystem, nb
 	witnesses := make([]witness.Witness, nbInstances)
 
 	for i := 0; i < nbInstances; i++ {
-		var assignment InnerCircuit
+		var assignment InnerCircuit1
 
 		var x, y fr_bls12377.Element
 		x.SetRandom()
@@ -142,144 +138,6 @@ func getProofsWitnesses(assert *test.Assert, ccs constraint.ConstraintSystem, nb
 	return proofs, witnesses
 }
 
-//------------------------------------------------------
-// outer circuit
-
-type BatchVerifyCircuit[FR emulated.FieldParams, G1El algebra.G1ElementT, G2El algebra.G2ElementT, GtEl algebra.GtElementT] struct {
-
-	// Number of proofs to batch
-	batchSizeProofs int
-
-	// dummy proofs, which are selected instead of the real proof, if the
-	// corresponding selector is 0. The dummy proofs always pass.
-	// TODO this should be a constant
-	DummyProof Proof[FR, G1El, G2El]
-
-	// proofs, verifying keys of the inner circuit
-	Proofs        []Proof[FR, G1El, G2El]
-	VerifyfingKey VerifyingKey[FR, G1El, G2El] // TODO this should be a constant
-
-	// selectors[i]==0/1 means that the i-th circuit is un/instantiated
-	// Selectors []frontend.Variable
-
-	// Corresponds to the public inputs of the inner circuit
-	PublicInners []Witness[FR]
-
-	// hash of the public inputs of the inner circuits
-	HashPub frontend.Variable `gnark:",public"`
-}
-
-func (circuit *BatchVerifyCircuit[FR, G1El, G2El, GtEl]) Define(api frontend.API) error {
-
-	// get Plonk verifier
-	curve, err := algebra.GetCurve[FR, G1El](api)
-	if err != nil {
-		return err
-	}
-
-	// check that hash(PublicInnters)==HashPub
-	var fr FR
-	h, err := recursion.NewHash(api, fr.Modulus(), true)
-	if err != nil {
-		return err
-	}
-	for i := 0; i < len(circuit.PublicInners); i++ {
-		for j := 0; j < len(circuit.PublicInners[i].Public); j++ {
-			toHash := curve.MarshalScalar(circuit.PublicInners[i].Public[j])
-			h.Write(toHash...)
-		}
-	}
-	s := h.Sum()
-	api.AssertIsEqual(s, circuit.HashPub)
-
-	// check that the proofs are correct
-	verifier, err := NewVerifier[FR, G1El, G2El, GtEl](api)
-	if err != nil {
-		return fmt.Errorf("new verifier: %w", err)
-	}
-	commitments := make([]gnark_kzg.Commitment[G1El], 2*circuit.batchSizeProofs)
-	proofs := make([]gnark_kzg.OpeningProof[FR, G1El], 2*circuit.batchSizeProofs)
-	points := make([]emulated.Element[FR], 2*circuit.batchSizeProofs)
-	for i := 0; i < circuit.batchSizeProofs; i++ {
-		//err = verifier.AssertProof(circuit.VerifyfingKey, circuit.Proofs[i], circuit.PublicInners[i])
-		commitmentPair, proofPair, pointPair, err := verifier.PrepareVerification(circuit.VerifyfingKey, circuit.Proofs[i], circuit.PublicInners[i])
-		if err != nil {
-			return err
-		}
-		copy(commitments[2*i:], commitmentPair)
-		copy(proofs[2*i:], proofPair)
-		copy(points[2*i:], pointPair)
-	}
-
-	kzgVerifier, err := gnark_kzg.NewVerifier[FR, G1El, G2El, GtEl](api)
-	if err != nil {
-		return err
-	}
-	err = kzgVerifier.BatchVerifyMultiPoints(commitments, proofs, points, circuit.VerifyfingKey.Kzg)
-
-	return err
-}
-
-func instantiateOuterCircuit[FR emulated.FieldParams, G1El algebra.G1ElementT, G2El algebra.G2ElementT, GtEl algebra.GtElementT](
-	assert *test.Assert,
-	batchSizeProofs int,
-	witnesses []witness.Witness,
-	innerCcs constraint.ConstraintSystem) BatchVerifyCircuit[FR, G1El, G2El, GtEl] {
-
-	// outer ciruit instantation
-	outerCircuit := BatchVerifyCircuit[FR, G1El, G2El, GtEl]{
-		PublicInners: make([]Witness[FR], batchSizeProofs),
-	}
-	for i := 0; i < len(witnesses); i++ {
-		outerCircuit.PublicInners[i] = PlaceholderWitness[FR](innerCcs)
-	}
-	outerCircuit.Proofs = make([]Proof[FR, G1El, G2El], batchSizeProofs)
-	for i := 0; i < batchSizeProofs; i++ {
-		outerCircuit.Proofs[i] = PlaceholderProof[FR, G1El, G2El](innerCcs)
-	}
-	outerCircuit.DummyProof = PlaceholderProof[FR, G1El, G2El](innerCcs)
-	outerCircuit.VerifyfingKey = PlaceholderVerifyingKey[FR, G1El, G2El](innerCcs)
-	outerCircuit.batchSizeProofs = batchSizeProofs
-	// outerCircuit.Selectors = make([]frontend.Variable, batchSizeProofs)
-
-	return outerCircuit
-}
-
-func assignWitness[FR emulated.FieldParams, G1El algebra.G1ElementT, G2El algebra.G2ElementT, GtEl algebra.GtElementT](
-	assert *test.Assert,
-	batchSizeProofs int,
-	frHashPub string,
-	witnesses []witness.Witness,
-	vk native_plonk.VerifyingKey,
-	proofs []native_plonk.Proof,
-	// selectors []int,
-) BatchVerifyCircuit[FR, G1El, G2El, GtEl] {
-
-	assignmentPubToPrivWitnesses := make([]Witness[FR], batchSizeProofs)
-	for i := 0; i < batchSizeProofs; i++ {
-		curWitness, err := ValueOfWitness[FR](witnesses[i])
-		assert.NoError(err)
-		assignmentPubToPrivWitnesses[i] = curWitness
-	}
-	assignmentVerifyingKeys, err := ValueOfVerifyingKey[FR, G1El, G2El](vk)
-	assert.NoError(err)
-	assignmentProofs := make([]Proof[FR, G1El, G2El], batchSizeProofs)
-	for i := 0; i < batchSizeProofs; i++ {
-		assignmentProofs[i], err = ValueOfProof[FR, G1El, G2El](proofs[i])
-		assert.NoError(err)
-	}
-	assignmentDummyProof, err := ValueOfProof[FR, G1El, G2El](proofs[0])
-	outerAssignment := BatchVerifyCircuit[FR, G1El, G2El, GtEl]{
-		Proofs:        assignmentProofs,
-		VerifyfingKey: assignmentVerifyingKeys,
-		PublicInners:  assignmentPubToPrivWitnesses,
-		HashPub:       frHashPub,
-		DummyProof:    assignmentDummyProof,
-	}
-
-	return outerAssignment
-}
-
 // set the outer proof
 func TestBatchVerify(t *testing.T) {
 
@@ -287,7 +145,7 @@ func TestBatchVerify(t *testing.T) {
 
 	// get ccs, vk, pk, srs
 	batchSizeProofs := 10
-	innerCcs, vk, pk, _ := GetInnerCircuitData()
+	innerCcs, vk, pk, _ := getInnerCircuitData()
 
 	// get tuples (proof, public_witness)
 	proofs, witnesses := getProofsWitnesses(assert, innerCcs, batchSizeProofs, pk, vk)
@@ -306,13 +164,8 @@ func TestBatchVerify(t *testing.T) {
 	var frHashPub fr_bw6761.Element
 	frHashPub.SetBytes(hashPub)
 
-	// selectors := make([]int, batchSizeProofs)
-	// for i := 0; i < batchSizeProofs; i++ {
-	// 	selectors[i] = i % 2
-	// }
-
 	// outer circuit
-	outerCircuit := instantiateOuterCircuit[
+	outerCircuit := InstantiateOuterCircuit[
 		sw_bls12377.ScalarField,
 		sw_bls12377.G1Affine,
 		sw_bls12377.G2Affine,
@@ -324,7 +177,7 @@ func TestBatchVerify(t *testing.T) {
 	)
 
 	// witness assignment
-	outerAssignment := assignWitness[sw_bls12377.ScalarField,
+	outerAssignment := AssignWitness[sw_bls12377.ScalarField,
 		sw_bls12377.G1Affine,
 		sw_bls12377.G2Affine,
 		sw_bls12377.GT](
@@ -362,12 +215,11 @@ func TestBatchVerify(t *testing.T) {
 	// err = native_plonk.Verify(proof, bw6vk, publicWitness)
 	// assert.NoError(err)
 
-	ccs, err := frontend.Compile(
-		ecc.BW6_761.ScalarField(),
-		scs.NewBuilder,
-		&outerCircuit)
-	assert.NoError(err)
-	fmt.Printf("nb constraints: %d\n", ccs.GetNbConstraints())
+	// ccs, err := frontend.Compile(
+	// 	ecc.BW6_761.ScalarField(),
+	// 	scs.NewBuilder,
+	// 	&outerCircuit)
+	// assert.NoError(err)
 
 	err = test.IsSolved(&outerCircuit, &outerAssignment, ecc.BW6_761.ScalarField())
 	assert.NoError(err)
